(PC-22883)[PRO] refactor: rename more bis to more
