+++ pro/src/pages/Home/Venues/Venue.tsx
@@ -18,7 +18,7 @@ import { useNewOfferCreationJourney } from 'hooks'
 import useActiveFeature from 'hooks/useActiveFeature'
 import useAnalytics from 'hooks/useAnalytics'
 import { CircleArrowIcon, NotificationErrorIcon } from 'icons'
-import { ReactComponent as IcoPlus } from 'icons/full-more-bis.svg'
+import { ReactComponent as IcoPlus } from 'icons/full-more.svg'
 import { ReactComponent as PenIcon } from 'icons/ico-pen-black.svg'
 import { ReactComponent as IcoPlusCircle } from 'icons/ico-plus-circle.svg'
 import { Button, ButtonLink } from 'ui-kit'
