+++ pro/src/screens/Offers/Offers.tsx
@@ -22,7 +22,7 @@ import { Audience } from 'core/shared'
 import getUserValidatedOfferersNamesAdapter from 'core/shared/adapters/getUserValidatedOfferersNamesAdapter'
 import { SelectOption } from 'custom_types/form'
 import useAnalytics from 'hooks/useAnalytics'
-import { ReactComponent as AddOfferSvg } from 'icons/full-more-bis.svg'
+import { ReactComponent as AddOfferSvg } from 'icons/full-more.svg'
 import { ReactComponent as LibraryIcon } from 'icons/library.svg'
 import { ReactComponent as strokeUserIcon } from 'icons/stroke-user.svg'
 import ActionsBar from 'pages/Offers/Offers/ActionsBar'
