+++ pro/src/screens/VenueForm/VenueFormScreen.tsx
@@ -27,7 +27,7 @@ import useAnalytics from 'hooks/useAnalytics'
 import useCurrentUser from 'hooks/useCurrentUser'
 import useNotification from 'hooks/useNotification'
 import { MailOutlineIcon } from 'icons'
-import { ReactComponent as AddOfferSvg } from 'icons/full-more-bis.svg'
+import { ReactComponent as AddOfferSvg } from 'icons/full-more.svg'
 import { Button, Title } from 'ui-kit'
 import { ButtonVariant } from 'ui-kit/Button/types'
