+++ pro/src/ui-kit/Icons/Icons.tsx
@@ -32,7 +32,7 @@ const fullIcons: IconListItem[] = [
   { src: 'icons/full-back.svg' },
   { src: 'icons/full-download.svg' }, // TODO 3 uses left inside <Button> and <ButtonLink>
   { src: 'icons/full-parameters.svg' }, // TODO 1 use left inside a <Button>
-  { src: 'icons/full-more-bis.svg' }, // TODO still in 2 ButtonLink and 1 Button
+  { src: 'icons/full-more.svg' }, // TODO still in 2 ButtonLink and 1 Button
   { src: 'icons/full-down.svg' }, // TODO 1 use left inside a <Button>
   { src: 'icons/full-up.svg' }, // TODO 1 use left inside a <Button>
   { src: 'icons/full-right.svg' },
